--- create-ami/src/main/java/jp/gr/java_conf/uzresk/aws/ope/ami/CreateAMI.java
@@ -1,398 +1,0 @@
-package jp.gr.java_conf.uzresk.aws.ope.ami;
-
-import java.io.IOException;
-import java.io.PrintWriter;
-import java.io.StringWriter;
-import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.Comparator;
-import java.util.List;
-import java.util.concurrent.Future;
-
-import com.amazonaws.ClientConfiguration;
-import com.amazonaws.auth.DefaultAWSCredentialsProviderChain;
-import com.amazonaws.regions.RegionUtils;
-import com.amazonaws.services.ec2.AmazonEC2;
-import com.amazonaws.services.ec2.AmazonEC2Async;
-import com.amazonaws.services.ec2.AmazonEC2AsyncClient;
-import com.amazonaws.services.ec2.model.BlockDeviceMapping;
-import com.amazonaws.services.ec2.model.CreateImageRequest;
-import com.amazonaws.services.ec2.model.CreateImageResult;
-import com.amazonaws.services.ec2.model.CreateTagsRequest;
-import com.amazonaws.services.ec2.model.DeregisterImageRequest;
-import com.amazonaws.services.ec2.model.DescribeImagesRequest;
-import com.amazonaws.services.ec2.model.DescribeImagesResult;
-import com.amazonaws.services.ec2.model.Filter;
-import com.amazonaws.services.ec2.model.Image;
-import com.amazonaws.services.ec2.model.ImageState;
-import com.amazonaws.services.ec2.model.Tag;
-import com.amazonaws.services.lambda.runtime.Context;
-import com.amazonaws.services.lambda.runtime.LambdaLogger;
-import com.amazonaws.services.sqs.AmazonSQSAsync;
-import com.amazonaws.services.sqs.AmazonSQSAsyncClient;
-import com.amazonaws.services.sqs.model.CreateQueueRequest;
-import com.amazonaws.services.sqs.model.Message;
-import com.amazonaws.services.sqs.model.ReceiveMessageRequest;
-import com.amazonaws.services.sqs.model.ReceiveMessageResult;
-import com.amazonaws.services.sqs.model.SendMessageRequest;
-import com.amazonaws.services.sqs.model.SendMessageResult;
-import com.fasterxml.jackson.databind.ObjectMapper;
-
-public class CreateAMI {
-
-	private ClientConfiguration cc;
-
-	public void request(CreateAMIRequest createAMIRequest, Context context) {
-
-		LambdaLogger logger = context.getLogger();
-
-		try {
-
-			if (createAMIRequest.getInstanceId() == null) {
-				throw new IllegalArgumentException("instance id is null.");
-			}
-			if (createAMIRequest.getAmiName() == null) {
-				throw new IllegalArgumentException("AMI Name is null.");
-			}
-
-			String createAMIRequestDate = LocalDateTime.now().format(DateTimeFormatter.ofPattern("yyyyMMddHHmm"));
-			createAMIRequest.setCreateAMIRequestDate(createAMIRequestDate);
-			createAMIRequest.setAmiName(createAMIRequest.getAmiName() + "-" + createAMIRequestDate);
-
-			String imageId = createAMI(createAMIRequest, context);
-			createAMIRequest.setImageId(imageId);
-
-			createQueueMessage(createAMIRequest, context);
-
-			logger.log("[SUCCESS][" + createAMIRequest.getInstanceId() + "][CreateAMIRequest]"
-					+ "Start request of the instance has completed successfully." + createAMIRequest);
-
-		} catch (Exception e) {
-			logger.log("[ERROR][" + createAMIRequest.getInstanceId() + "][CreateAMIRequest] message[" + e.getMessage()
-					+ "] stackTrace[" + getStackTrace(e) + "]" + createAMIRequest);
-		}
-	}
-
-	String createAMI(CreateAMIRequest createAMIRequest, Context context) {
-
-		LambdaLogger logger = context.getLogger();
-		AmazonEC2Async client = createEC2Client();
-
-		String imageId = null;
-		try {
-			Future<CreateImageResult> result = client.createImageAsync(
-					new CreateImageRequest(createAMIRequest.getInstanceId(), createAMIRequest.getAmiName())
-							.withNoReboot(createAMIRequest.isNoReboot()));
-			while (!result.isDone()) {
-				Thread.sleep(1000);
-			}
-			imageId = result.get().getImageId();
-
-			logger.log("AMI Create Request End. instanceId[" + createAMIRequest.getInstanceId() + "] noReboot["
-					+ createAMIRequest.isNoReboot() + "] imageId[" + imageId + "]");
-		} catch (Exception e) {
-			throw new RuntimeException("An unexpected error at the time of AMI creation has occurred", e);
-		} finally {
-			client.shutdown();
-		}
-		return imageId;
-	}
-
-	SendMessageResult createQueueMessage(CreateAMIRequest createAMIRequest, Context context) {
-
-		LambdaLogger logger = context.getLogger();
-
-		final String queueName = createAMIRequest.getQueueName();
-		final String sqsEndpoint = createAMIRequest.getSqsEndpoint();
-
-		if (queueName == null || sqsEndpoint == null) {
-			logger.log("skip create queue. [" + createAMIRequest + "]");
-			return null;
-		}
-
-		AmazonSQSAsync client = createSQSClient();
-		client.setEndpoint(sqsEndpoint);
-		createAMIRequest.setSendMessageTimeMillis(System.currentTimeMillis());
-
-		try {
-			CreateQueueRequest req = new CreateQueueRequest(queueName);
-			String queueUrl = client.createQueue(req).getQueueUrl();
-
-			SendMessageRequest sendMessage = new SendMessageRequest();
-			sendMessage.setQueueUrl(queueUrl);
-			ObjectMapper om = new ObjectMapper();
-			sendMessage.setMessageBody(om.writeValueAsString(createAMIRequest));
-
-			Future<SendMessageResult> result = client.sendMessageAsync(sendMessage);
-			while (!result.isDone()) {
-				Thread.sleep(100);
-			}
-			return result.get();
-
-		} catch (Exception e) {
-			throw new RuntimeException("unexpected error occured in the create queue request.", e);
-		} finally {
-			client.shutdown();
-		}
-	}
-
-	public void checkStateAndPargeAMI(CheckAMIStateRequest checkAMIStateRequest, Context context) {
-
-		LambdaLogger logger = context.getLogger();
-
-		final String queueName = checkAMIStateRequest.getQueueName();
-		final String sqsEndpoint = checkAMIStateRequest.getSqsEndpoint();
-
-		// illegal parameter
-		if (queueName == null || sqsEndpoint == null) {
-			logger.log("[ERROR][checkStateAndPargeAMI]QueueName or SQSEndpoint is not found Parameter. ");
-			throw new IllegalArgumentException(
-					"QueueName or SQSEndpoint is not found Parameter. " + "[" + checkAMIStateRequest + "]");
-		}
-
-		// Only the specified number, reliably acquired
-		int numberOfMessages = checkAMIStateRequest.getNumberOfMessages();
-		for (int i = 0; i < numberOfMessages; i++) {
-			AmazonSQSAsync client = createSQSClient();
-			client.setEndpoint(sqsEndpoint);
-			try {
-				String queueUrl = client.createQueue(queueName).getQueueUrl();
-
-				ReceiveMessageRequest req = new ReceiveMessageRequest(queueUrl).withVisibilityTimeout(5)
-						.withMaxNumberOfMessages(checkAMIStateRequest.getNumberOfMessages());
-				Future<ReceiveMessageResult> result = client.receiveMessageAsync(req);
-				while (!result.isDone()) {
-					Thread.sleep(100);
-				}
-				result.get().getMessages().stream()
-						.forEach(s -> checkStateAndPargeAMI(s, checkAMIStateRequest, context));
-
-			} catch (Exception e) {
-				logger.log("[ERROR][checkStateAndPargeAMI] message[" + e.getMessage() + "] stackTrace["
-						+ getStackTrace(e) + "] [" + checkAMIStateRequest + "]");
-
-			} finally {
-				client.shutdown();
-			}
-		}
-	}
-
-	void checkStateAndPargeAMI(Message message, CheckAMIStateRequest checkAMIStateRequest, Context context) {
-
-		LambdaLogger logger = context.getLogger();
-
-		ObjectMapper om = new ObjectMapper();
-		CreateAMIRequest createAMIRequest;
-		try {
-			createAMIRequest = om.readValue(message.getBody(), CreateAMIRequest.class);
-		} catch (IOException e) {
-			deleteQueueMessage(message, checkAMIStateRequest, context);
-			throw new RuntimeException("SQS message could not be parsed");
-		}
-
-		long sendMessageTime = createAMIRequest.getSendMessageTimeMillis()
-				+ 1000 * createAMIRequest.getAmiCreatedCheckTimeoutSec();
-		long now = System.currentTimeMillis();
-
-		// Status check of the instance has timed out or not
-		if (sendMessageTime < now) {
-			deleteQueueMessage(message, checkAMIStateRequest, context);
-			throw new RuntimeException("Status check of the instance has timed out. " + createAMIRequest);
-		}
-
-		AmazonEC2Async client = createEC2Client();
-		try {
-			String imageState = getImageState(client, createAMIRequest, context);
-			if (ImageState.Available.toString().equals(imageState)) {
-				createImageTags(client, createAMIRequest, context);
-				pargeImage(client, createAMIRequest, context);
-				deleteQueueMessage(message, checkAMIStateRequest, context);
-			}
-			logger.log("[SUCCESS][" + createAMIRequest.getInstanceId() + "] "
-					+ "Creation of AMI, additional tags, generation management has completed successfully. ["
-					+ createAMIRequest + "]");
-		} finally {
-			client.shutdown();
-		}
-
-	}
-
-	void deleteQueueMessage(Message message, CheckAMIStateRequest checkAMIStateRequest, Context context) {
-
-		AmazonSQSAsync client = createSQSClient();
-		client.setEndpoint(checkAMIStateRequest.getSqsEndpoint());
-		try {
-			String queueUrl = client.createQueue(checkAMIStateRequest.getQueueName()).getQueueUrl();
-			client.deleteMessage(queueUrl, message.getReceiptHandle());
-		} catch (Exception e) {
-			final String msg = "can not delete message. [" + checkAMIStateRequest + "]";
-			throw new RuntimeException(msg, e);
-		} finally {
-			client.shutdown();
-		}
-	}
-
-	void createImageTags(AmazonEC2Async client, CreateAMIRequest createAMIRequest, Context context) {
-
-		// LambdaLogger logger = context.getLogger();
-		String instanceId = createAMIRequest.getInstanceId();
-		String imageId = createAMIRequest.getImageId();
-
-		List<Tag> tags = new ArrayList<Tag>();
-		tags.add(new Tag("InstanceId", instanceId));
-		tags.add(new Tag("BackupType", "auto"));
-		String requestDate = LocalDateTime.now().format(DateTimeFormatter.ofPattern("yyyyMMddHHmm"));
-		tags.add(new Tag("RequestDate", requestDate));
-
-		// AMIにタグを付与する。
-		CreateTagsRequest createTagsRequest = new CreateTagsRequest().withResources(imageId);
-		createTagsRequest.setTags(tags);
-		client.createTags(createTagsRequest);
-
-		// EBS snapshotにタグを付与する
-		tags.add(new Tag("ImageId", imageId)); // snapshotにはimageIdを付けておく。
-
-		List<String> snapshotIds = getSnapshotIdsFromImageId(client, createAMIRequest, context);
-		CreateTagsRequest snapshotTagsRequest = new CreateTagsRequest().withResources(snapshotIds);
-
-		snapshotTagsRequest.setTags(tags);
-		client.createTags(snapshotTagsRequest);
-	}
-
-	// AMIに紐づくEBSSnapshotIdを取得します。
-	List<String> getSnapshotIdsFromImageId(AmazonEC2Async client, CreateAMIRequest request, Context context) {
-
-		// LambdaLogger logger = context.getLogger();
-		String imageId = request.getImageId();
-
-		DescribeImagesResult result = client.describeImages(new DescribeImagesRequest().withImageIds(imageId));
-
-		List<String> snapshotIds = new ArrayList<String>();
-
-		for (Image image : result.getImages()) {
-			for (BlockDeviceMapping block : image.getBlockDeviceMappings()) {
-				snapshotIds.add(block.getEbs().getSnapshotId());
-			}
-		}
-		return snapshotIds;
-	}
-
-	String getImageState(AmazonEC2Async client, CreateAMIRequest createAMIRequest, Context context) {
-		// LambdaLogger logger = context.getLogger();
-		String imageId = createAMIRequest.getImageId();
-
-		DescribeImagesRequest req = new DescribeImagesRequest().withImageIds(imageId);
-		DescribeImagesResult result = client.describeImages(req);
-		return result.getImages().get(0).getState();
-	}
-
-	void pargeImage(AmazonEC2Async client, CreateAMIRequest request, Context context) {
-
-		LambdaLogger logger = context.getLogger();
-		String instanceId = request.getInstanceId();
-		int generationCount = request.getGenerationCount();
-
-		// InstanceIdというタグ名でAMIを取得します。
-		Filter filter = new Filter();
-		filter.withName("tag:InstanceId").withValues(instanceId);
-		DescribeImagesRequest describeImageRequest = new DescribeImagesRequest().withFilters(filter);
-		DescribeImagesResult result = client.describeImages(describeImageRequest);
-
-		// RequestDateの昇順で並べ替えます。
-		List<Image> images = result.getImages();
-		Collections.sort(images, new ImagesComparator());
-
-		// debug
-		logger.log("image size." + images.size());
-		for (Image image : images) {
-			logger.log(image.getTags().toString());
-		}
-
-		// 世代管理保持数 < snapshotの数の場合、対象をpargeします。
-		List<String> pargeList = new ArrayList<>();
-		int imagesSize = images.size();
-		if (generationCount < imagesSize) {
-			for (int i = 0; i < imagesSize - generationCount; i++) {
-				Image image = images.get(i);
-				// （念のため）availableのものだけparge対象とする。
-				if (ImageState.Available.toString().equals(image.getState())) {
-					pargeImages(client, image);
-					String imageId = image.getImageId();
-					logger.log("parge image. imageId[" + imageId + "] instanceId[" + instanceId + "] generationCount["
-							+ generationCount + "]");
-					pargeList.add(imageId);
-				}
-			}
-		}
-	}
-
-	private class ImagesComparator implements Comparator<Image> {
-		@Override
-		public int compare(Image o1, Image o2) {
-			long o1RequestDate = 0;
-			for (Tag tag : o1.getTags()) {
-				if ("RequestDate".equals(tag.getKey())) {
-					o1RequestDate = Long.parseLong(tag.getValue());
-				}
-			}
-			long o2RequestDate = 0;
-			for (Tag tag : o2.getTags()) {
-				if ("RequestDate".equals(tag.getKey())) {
-					o2RequestDate = Long.parseLong(tag.getValue());
-				}
-			}
-			if (o1RequestDate - o2RequestDate > 0) {
-				return 1;
-			} else if (o1RequestDate - o2RequestDate < 0) {
-				return -1;
-			} else {
-				return 0;
-			}
-		}
-	}
-
-	// 指定されたAMIとそれに紐づくEBSSnapshotをPargeします。
-	private void pargeImages(AmazonEC2 ec2, Image image) {
-		String imageId = image.getImageId();
-		ec2.deregisterImage(new DeregisterImageRequest(imageId));
-		// for (BlockDeviceMapping block : image.getBlockDeviceMappings()) {
-		// String snapshotId = block.getEbs().getSnapshotId();
-		// ec2.deleteSnapshot(new
-		// DeleteSnapshotRequest().withSnapshotId(snapshotId));
-		// }
-	}
-
-	protected AmazonEC2Async createEC2Client() {
-		String regionName = System.getenv("AWS_DEFAULT_REGION");
-
-		return RegionUtils.getRegion(regionName).createClient(AmazonEC2AsyncClient.class,
-				new DefaultAWSCredentialsProviderChain(), cc);
-	}
-
-	protected AmazonSQSAsync createSQSClient() {
-		String regionName = System.getenv("AWS_DEFAULT_REGION");
-		return RegionUtils.getRegion(regionName).createClient(AmazonSQSAsyncClient.class,
-				new DefaultAWSCredentialsProviderChain(), cc);
-	}
-
-	void setClientConfiguration(ClientConfiguration cc) {
-		this.cc = cc;
-	}
-
-	ClientConfiguration getClientConfiguration() {
-		if (this.cc == null) {
-			return new ClientConfiguration();
-		}
-		return this.cc;
-	}
-
-	protected String getStackTrace(Exception e) {
-		StringWriter sw = new StringWriter();
-		PrintWriter pw = new PrintWriter(sw);
-		e.printStackTrace(pw);
-		pw.flush();
-		return sw.toString();
-	}
-}
