+++ create-image/src/main/java/jp/gr/java_conf/uzresk/aws/ope/image/ImageCreateRequests.java
@@ -1,0 +1,11 @@
+package jp.gr.java_conf.uzresk.aws.ope.image;
+
+import java.util.List;
+
+import lombok.Data;
+
+@Data
+public class ImageCreateRequests {
+
+	private List<ImageCreateRequest> imageCreateRequests;
+}
