+++ create-image/src/test/java/jp/gr/java_conf/uzresk/aws/ope/TestContext.java
